refactor(util): name the loopback predicate for the guarantee it makes
`endpoint_is_loopback` lived in the rclone importer, where it decides whether a self-signed certificate is acceptable on export. It is about to get a second caller with a different question, so it moves to `util::endpoint` as `endpoint_stays_on_this_machine`: the name states what a caller depends on rather than what the string looks like, which matters here because a similarly named predicate sits nearby and promises something else.

`providers::s3::is_local_s3_endpoint` also accepts `*.local` and `*.localhost`. That is right for its job, since local bridges publish mDNS names and self-sign their certificates, and wrong for confidentiality, since a `.local` host is reached across a network segment other machines share. Both functions now carry a note saying why they are not the same predicate, and the moved one gains a test that asserts mDNS and LAN addresses are NOT this machine. That test is the one that fails if someone later merges the two for tidiness.

The seven existing cases travel with it, including the two that matter most: an IPv6 literal in brackets, and a path segment named `localhost` that must not be read as the host.
